Add local fork upkeep tools
Note to self: look into reporting the need for the FixVCSoln thing to someone; that shouldn't be happening in the first place. Use case: adding a new .cpp file to the Visual Studio project in one branch, then switching to another branch, will make VS think the file is still there in *only some* ways. Leads to a whole bunch of headaches -- the only solution is to nuke, and redownload a clean copy of, the entire VS folder.